Big smoothings and lots of questions! :)
* Be careful with this phrase:

"if the **PCBs** don't work as your expect"

What that actually says is "If you design your circuit or PCB so that you have wired up the battery or the transistors the wrong way round or you have shorted the opamp output and so your circuit doesn't work as you expected, that's OK, we'll refund your money."

That is the user making a mistake or just being careless in checking. We must not offer a refund for that. 

The sort of things we can refund a user's money for is if, for example, our continuity check fails to detect a fault or if the masks are misaligned or maybe if we fail to spot and correct silkscreen over the pads. Or if the PCB laminate has faults in it or we ship the wrong solder mask colour.

I have changed this phrase to:

"if you are not satisfied with the quality of our PCBs,"


* You had a section for "PCB capabilities" and then almost the same thing again in "Manufacturing Specifications".

Then another section called "Manufacturing Specifications".

I have edited the section on "PCB capabilities" and deleted the 1st section on "Manufacturing Specifications".

* Note that you state minimum PCB dimensions = 2cm in "PCB capabilities" and = 5cm in "Manufacturing Specifications": which is correct???


* In the "Manufacturing Specifications", you have thickness tolerances for >=1mm of +/-10% but for <=1mm of +/-0.1%. 

Is this correct? Is it not >=1mm of +/-10% and <=1mm of +/-0.1mm (and similarly for the mil dimensions?)


* What is "Insulation layer thickness" given as 0.075mm - 5.0mm? You have already given "Available Board Thickness" as 0.6mm (except 4 layer) - 2.0mm.


* "Minimum inner trace width (for 4 layer)" and "Minimum inner trace/vias/pads space (for 4 layer)".

Is this the same for 6 layers or more? Should these say: "...(for >2 layers)"?


* "Outer Layer Copper Thickness" Is this really fixed to a single thickness?


* "Aspect Ratio" What is this for? PCB or pads or elements drawn in the silkscreen?